+++ packages/codegen/src/sdk/generators/index.ts
@@ -1,0 +1,6 @@
+import { NodePlopAPI } from 'plop';
+import SdkReadmeGenerator from './readme';
+
+export default function setNextGenerators(plop: NodePlopAPI) {
+  SdkReadmeGenerator(plop);
+}