+++ packages/next/src/hooks/auth/useAuthRequestChallengeSolana.ts
@@ -1,40 +1,35 @@
 import {
   requestChallengeSolanaOperation as operation,
   RequestChallengeSolanaRequest,
-  RequestChallengeSolanaResponse,
 } from '@moralisweb3/common-auth-utils';
-import { fetcher } from '../../utils/fetcher';
 import { FetchParams } from '../types';
-import { useCallback } from 'react';
-import useSWR from 'swr';
+import { useResolver } from '../resolvers';
 
 export type RequestChallengeSolanaRequestClient = Pick<RequestChallengeSolanaRequest, 'address' | 'network'>;
 
 export const useAuthRequestChallengeSolana = (
-  request: RequestChallengeSolanaRequestClient,
+  request?: RequestChallengeSolanaRequestClient,
   fetchParams?: FetchParams,
 ) => {
-  const endpoint = 'auth/requestChallengeSolana';
-
-  const { data, error, isValidating, mutate } = useSWR<RequestChallengeSolanaResponse>(
-    [endpoint, { operation, request }],
-    fetcher,
-    {
-      revalidateOnMount: request ? true : false,
-      revalidateOnFocus: false,
-      ...fetchParams,
-    },
-  );
-
-  const requestChallengeAsync = useCallback((params: RequestChallengeSolanaRequestClient) => {
-    return mutate(fetcher(endpoint, { operation, request: params }));
-  }, []);
+  const { data, error, fetch, isFetching } = useResolver({
+    endpoint: 'auth/requestChallengeSolana',
+    operation,
+    request,
+    fetchParams,
+  });
 
   return {
     challenge: data,
     error,
-    requestChallengeAsync,
-    refetch: async () => mutate(),
-    isValidating,
+    requestChallengeAsync: fetch,
+    /**
+     * @deprecated use `fetch()` instead
+     */
+    refetch: () => fetch(),
+    isFetching,
+    /**
+     * @deprecated use `isFetching` instead
+     */
+    isValidating: isFetching,
   };
 };