+++ packages/solUtils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moralisweb3/sol-utils
 
+## 2.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-sol-utils@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-sol-utils@2.8.1
+
 ## 2.8.0
 
 ### Minor Changes